Some additional refactoring to clean up the big Marathon.jsx.
